Finalize Dto implementation

+++ src/Capitan360.Application/Features/Companies/CompanyCommissionses/Commands/Create/CreateCompanyCommissionsCommand.cs
@@ -2,9 +2,9 @@
 
 public record CreateCompanyCommissionsCommand(
     int CompanyId,
-    decimal CommissionFromCaptainCargoWebSite,
-    decimal CommissionFromCompanyWebSite,
-    decimal CommissionFromCaptainCargoWebService,
-    decimal CommissionFromCompanyWebService,
-    decimal CommissionFromCaptainCargoPanel,
-    decimal CommissionFromCaptainCargoDesktop);
+    long CommissionFromCaptainCargoWebSite,
+    long CommissionFromCompanyWebSite,
+    long CommissionFromCaptainCargoWebService,
+    long CommissionFromCompanyWebService,
+    long CommissionFromCaptainCargoPanel,
+    long CommissionFromCaptainCargoDesktop);
